+++ build_mini.sh
@@ -1,3 +1,7 @@
+if [[ -n "${IDF_PATH}" ]]; then
+  echo "Idf env not set! call idf.py"
+  exit 1
+fi
 rm -Rf target .embuild
 cargo build -p native_code --config 'env.ln_board="mini"' &&
   cargo build -p extra_code --config 'env.ln_board="mini"' &&
